new improved Merkle proofs
