+++ chain/tests/test_txhashset_raw_txs.rs
@@ -25,8 +25,8 @@ use chain::txhashset;
 use chain::txhashset::TxHashSet;
 use chain::types::Tip;
 use chain::ChainStore;
-use core::core::pmmr::MerkleProof;
 use core::core::target::Difficulty;
+use core::core::MerkleProof;
 use core::core::{Block, BlockHeader};
 use keychain::Keychain;
 use wallet::libtx::{build, reward};
