+++ wallet/src/libtx/slate.rs
@@ -17,8 +17,7 @@
 use rand::thread_rng;
 use uuid::Uuid;
 
-use core::core::committed;
-use core::core::committed::Committed;
+use core::core::Committed;
 use core::core::{amount_to_hr_string, Transaction};
 use keychain::{BlindSum, BlindingFactor, Keychain};
 use libtx::error::{Error, ErrorKind};
